fix: add nil checks in WeatherMapService

+++ app/services/weather_map_service.rb
@@ -1,16 +1,15 @@
 # frozen_string_literal: true
 
 class WeatherMapService
-  attr_accessor :uri
+  attr_accessor :lat, :lon, :api_key, :uri
 
   def initialize
     return if Rails.application.credentials.openweathermap.blank?
 
-    lat = Rails.application.credentials.openweathermap[:lat]
-    lon = Rails.application.credentials.openweathermap[:lon]
-    api_key = Rails.application.credentials.openweathermap[:api_key]
     base_url = "https://api.openweathermap.org"
-
+    @lat = Rails.application.credentials.openweathermap[:lat]
+    @lon = Rails.application.credentials.openweathermap[:lon]
+    @api_key = Rails.application.credentials.openweathermap[:api_key]
     @uri = "#{base_url}/data/2.5/onecall?lat=#{lat}&lon=#{lon}&appid=#{api_key}&units=metric&exclude=minutely&lang=de"
   end
 
@@ -19,13 +18,17 @@ class WeatherMapService
   #
   # @return [OpenWeatherMap] latest created OpenWeatherMap object
   def import
+    return nil if lat.blank? || lon.blank? || api_key.blank?
+
     request_service = ApiRequestService.new(uri)
     result = request_service.get_request
 
+    return nil if result.blank?
     return nil unless result.code == "200"
     return nil if result.body.blank?
 
     weather_map = nil
+
     OpenWeatherMap.transaction do
       OpenWeatherMap.destroy_all
       weather_map = OpenWeatherMap.create(data: JSON.parse(result.body))
